Added AnimatorComponent + Implementation

+++ THE_DEV_GAME/PROYECT/THE DEV GAME/THE DEV GAME/Vector2D.cpp
@@ -51,7 +51,7 @@ Vector2D& Vector2D::operator/=(const Vector2D& vec) {
 	return this->Divide(vec);
 }
 
-Vector2D& Vector2D::operator*(const int& i) {
+Vector2D& Vector2D::operator*(const float& i) {
 	this->x *= i;
 	this->y *= i;
 	return *this;
